Feature: user_param
Add an opaque `user_param` pointer to the `mp3_options` struct so that it is accessible to user callback functions.  It will typically be used to point to an enclosing C++ object.

Feature: user_param

Add an opaque `user_param` pointer to the `mp3_options` struct so that it is accessible to user callback functions.  It will typically be used to point to an enclosing C++ object.

Update mp32pcm.h